Fixes for fightplan download from lenny
The sequence is now: Download one of the existing flightplans. Check that this is the one you are looking for. Click on Open FP button. The flightplan is ACTIVE now and is transmitted to the other ATCs using the Flightplan Exchange Server. This status change is also done at the flightgear-atc server, so this FP cannot be downloaded by any other ATC.
Before the OpenFP click the FP is still in FILED mode, so you can select again, if this was the wrong Flightplan.